+++ packages/ssx-react/package.json
@@ -53,8 +53,7 @@
   "private": false,
   "scripts": {
     "build": "tsup",
-    "watch": "tsup --watch" 
-
+    "watch": "tsup --watch"
   },
   "dependencies": {
     "@spruceid/ssx": "1.1.1",
