--- packages/ssx-react/src/common.ts
@@ -1,28 +1,0 @@
-import { ReactNode } from 'react';
-import type { SSX, SSXClientConfig } from '@spruceid/ssx';
-
-/** Interface for SSX Web3 Provider. */
-export interface SSXWeb3Provider {
-  /** web3 Provider. */
-  provider: any;
-  /** web3 Provider Loaded. */
-  providerLoaded?: boolean;
-}
-
-/** Props for SSX Provider. */
-export interface SSXProviderProps {
-  /** Optional SSX configuration, used for instantiating an SSX Instance. */
-  ssxConfig?: SSXClientConfig;
-  /** Provider child nodes, for rendering. */
-  children: ReactNode;
-  /** Optional SSX Signer. Will . */
-  web3Provider?: SSXWeb3Provider;
-}
-
-/** Interface for contents provided to the Hook. */
-export interface SSXContextInterface {
-  /** SSX Instance. */
-  ssx: SSX | undefined;
-  /** SSX Instance loading state. */
-  ssxLoaded: boolean;
-}
