+++ packages/ssx-react/tsup.config.ts
@@ -10,4 +10,5 @@ export default defineConfig({
       jsx: 'react-jsx',
     },
   },
+  external: ['ssx-wagmi'],
 });
